Correcciones y errores pendientes (LEER DESCRIPC)
AGREGUE UNA PLANILLA con errores, ahi vayan agregando todo lo que vayan
encontrando, y si lo pudieron solucionar, es la unica forma de tener un
seguimiento claro,y que no se nos pase nada por alto.por otro lado cada
error que corrijamos puede llegar a hacer que otra cosa deje de
funcionar (espero que no) por eso mismo SEAMOS ORDENADOS y agreguemos
los errores que vayamos encontrando/corrigiendo ahi.

+++ TP/src/PagoElectronico/ABMs/ABM Cliente/AltaCliente.cs
@@ -94,7 +94,7 @@ namespace PagoElectronico.ABMs.ABM_Cliente
             }
 
             GestorDeSistema.crearCliente(NombreClitextBox.Text, ApellidoNuevoClitextBox.Text, Convert.ToInt32(TipoDcomboBox.SelectedValue.ToString()),
-                                         Convert.ToInt32(NroDocClitextBox.Text), MailtextBox.Text, Convert.ToInt32(PaiscomboBox.SelectedValue.ToString()), DomicilioCalletextBox.Text,
+                                         Convert.ToDecimal(NroDocClitextBox.Text), MailtextBox.Text, Convert.ToDecimal(PaiscomboBox.SelectedValue.ToString()), DomicilioCalletextBox.Text,
                                          DomNumerotextBox.Text, DomicilioPisoTextBox.Text, DomicilioDeptotextBox.Text,
                                          LocalidadtextBox.Text, NacionalidadtextBox.Text, fechaNacimientodateTimePicker.Value);
             int clienteID = GestorDeSistema.obtenerClienteRecienCreado();
